requirements.txt
langchain>=0.2.0
langchain-community>=0.2.0
langchain-core>=0.2.0
langchain-huggingface>=0.0.3
langchain-text-splitters>=0.2.0
faiss-cpu>=1.8.0
sentence-transformers>=2.7.0
transformers>=4.41.0
accelerate>=0.30.0
pymupdf>=1.24.0
python-docx>=1.1.0
gradio>=4.36.0
